refactor: 로직에 따라 클래스 분리
- 클래스가 한 가지 일만 담당하게 하기 위해 당첨 결과를 비교하는 로직과 보너스 번호를 비교하는 로직을 분리합니다.
